requirement.txt
absl-py==2.1.0
addict==2.4.0
albumentations==1.3.0
annotated-types==0.6.0
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
audeer==2.0.0
audiofile==1.4.0
audioread==3.0.1
audmath==1.4.0
audtorch==0.6.4
Babel==2.13.1
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.12.2
bleach==6.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.0
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.6
debugpy==1.8.0
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
ever-beta==0.2.3
examples==0.1.2
exceptiongroup==1.2.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema==2.19.0
flatbuffers==24.3.25
fonttools==4.45.1
fqdn==1.5.1
gast==0.4.0
google-auth==2.28.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.0
h5py==3.11.0
idna==3.6
imageio==2.33.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.27.1
ipython @ file:///croot/ipython_1691532092695/work
ipywidgets==8.1.1
isoduration==20.11.0
jedi @ file:///tmp/build/80754af9/jedi_1644315233700/work
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==18.1.1
librosa==0.10.1
llvmlite==0.41.1
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mistune==3.0.2
mmcv-full==1.7.1
msgpack==1.0.7
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
ninja==1.11.1.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.2.0.32
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
pooch==1.8.1
prettytable==3.10.0
prometheus-client==0.19.0
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
protobuf==4.25.3
psutil==5.9.6
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==0.32.2
pydantic_core==2.18.2
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
qudida==0.0.4
referencing==0.31.1
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
-e git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588#egg=segment_anything
Send2Trash==1.8.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tabulate==0.9.0
termcolor==2.3.0
terminado==0.18.0
threadpoolctl==3.3.0
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
tqdm==4.66.1
traitlets @ file:///croot/traitlets_1671143879854/work
ttach==0.0.3
types-python-dateutil==2.8.19.14
typing_extensions==4.5.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.1.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
yacs==0.1.8
yapf==0.40.2
zipp==3.17.0
